+++ RemoteAgents/AgentSatisfactionSupply.cs
@@ -9,7 +9,7 @@ using LlamaLibrary.RemoteWindows;
 
 namespace LlamaLibrary.RemoteAgents
 {
-    public class AgentSatisfactionSupply: AgentInterface<AgentSatisfactionSupply>, IAgent
+    public class AgentSatisfactionSupply : AgentInterface<AgentSatisfactionSupply>, IAgent
     {
         public IntPtr RegisteredVtable => Offsets.VTable;
         private static class Offsets
@@ -27,6 +27,7 @@ namespace LlamaLibrary.RemoteAgents
             [Offset("Search 0F B7 73 ? BA ? ? ? ? E8 ? ? ? ? Add 3 Read8")]
             internal static int MaxRep;
             [Offset("Search 8B 53 ? 4C 8D 43 ? 48 8B CB Add 2 Read8")]
+
             //[OffsetCN("Search 8B 53 ? 48 8B CB E8 ? ? ? ? BA ? ? ? ? Add 2 Read8")]
             internal static int Npc;
             [Offset("Search 89 43 ? E8 ? ? ? ? 48 8B D0 48 8D 4D ? E8 ? ? ? ? Add 2 Read8")]
@@ -36,13 +37,13 @@ namespace LlamaLibrary.RemoteAgents
             [Offset("Search 48 89 5C 24 ? 48 89 6C 24 ? 48 89 74 24 ? 57 48 83 EC ? 48 8B 01 41 0F B6 E9 41 8B F8")]
             internal static IntPtr OpenWindow;
             [Offset("Search 4C 8D 43 ? 48 8B CB E8 ? ? ? ? BA ? ? ? ? Add 3 Read8")]
+
             //[OffsetCN("Search 4C 8D 43 ? 8B 53 ? 48 8B CB Add 3 Read8")]
             internal static int NpcId;
-
         }
+
         protected AgentSatisfactionSupply(IntPtr pointer) : base(pointer)
         {
-
         }
 
         public byte DeliveriesRemaining => Core.Memory.Read<byte>(Pointer + Offsets.DeliveriesRemaining);
@@ -70,11 +71,12 @@ namespace LlamaLibrary.RemoteAgents
                 await Coroutine.Wait(5000, () => !SatisfactionSupply.Instance.IsOpen);
             }
 
-            Core.Memory.CallInjected64<IntPtr>(Offsets.OpenWindow,
+            Core.Memory.CallInjected64<IntPtr>(
+                Offsets.OpenWindow,
                                                Pointer,
-                                               (uint)0,
-                                               (npc),
-                                               (uint)1);
+0U,
+                                               npc,
+1U);
 
             await Coroutine.Wait(5000, () => SatisfactionSupply.Instance.IsOpen);
